requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-jsonfield==1.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
